// apps/api/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  OWNER: 'OWNER',
  MANAGER: 'MANAGER',
  CASHIER: 'CASHIER',
  INVENTORY_CLERK: 'INVENTORY_CLERK'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const SaleStatus = {
  OPEN: 'OPEN',
  COMPLETED: 'COMPLETED',
  VOIDED: 'VOIDED',
  PARTIALLY_REFUNDED: 'PARTIALLY_REFUNDED',
  REFUNDED: 'REFUNDED'
} as const

export type SaleStatus = (typeof SaleStatus)[keyof typeof SaleStatus]


export const PaymentMethod = {
  CASH: 'CASH',
  CARD: 'CARD',
  OTHER: 'OTHER'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const InventoryMovementType = {
  SALE: 'SALE',
  SALE_VOID: 'SALE_VOID',
  REFUND: 'REFUND',
  PURCHASE_RECEIVED: 'PURCHASE_RECEIVED',
  DAMAGE: 'DAMAGE',
  EXPIRED: 'EXPIRED',
  THEFT: 'THEFT',
  MANUAL_ADJUSTMENT: 'MANUAL_ADJUSTMENT'
} as const

export type InventoryMovementType = (typeof InventoryMovementType)[keyof typeof InventoryMovementType]


export const ShiftStatus = {
  OPEN: 'OPEN',
  CLOSED: 'CLOSED',
  FORCE_CLOSED: 'FORCE_CLOSED'
} as const

export type ShiftStatus = (typeof ShiftStatus)[keyof typeof ShiftStatus]


export const PurchaseOrderStatus = {
  DRAFT: 'DRAFT',
  ORDERED: 'ORDERED',
  PARTIALLY_RECEIVED: 'PARTIALLY_RECEIVED',
  RECEIVED: 'RECEIVED',
  CANCELLED: 'CANCELLED'
} as const

export type PurchaseOrderStatus = (typeof PurchaseOrderStatus)[keyof typeof PurchaseOrderStatus]